Fix GCC warning cleanup

+++ benchmarks/db_protocol_synthetic_bench.cxx
@@ -241,7 +241,7 @@ void consume_wire_text(
 				}
 				acc += v;
 			} else {
-				acc += static_cast<std::uint64_t>(sv.size());
+				acc += sv.size();
 			}
 			p += field_len;
 		}
